+++ core/event-service/src/main/java/ru/practicum/categories/service/CategoryService.java
@@ -1,7 +1,7 @@
 package ru.practicum.categories.service;
 
-import ru.practicum.categories.dto.CategoryDto;
-import ru.practicum.categories.dto.NewCategoryDto;
+import ru.practicum.dto.categories.CategoryDto;
+import ru.practicum.dto.categories.NewCategoryDto;
 
 import java.util.List;
 
@@ -9,11 +9,11 @@ public interface CategoryService  {
 
     CategoryDto createCategory(NewCategoryDto newCategoryDto);
 
-    CategoryDto updateCategory(Integer catId, CategoryDto categoryDto);
+    CategoryDto updateCategory(Long catId, CategoryDto categoryDto);
 
-    void deleteCategory(Integer catId);
+    void deleteCategory(Long catId);
 
-    CategoryDto getCategory(Integer catId);
+    CategoryDto getCategory(Long catId);
 
     List<CategoryDto> getCategories(Integer from, Integer size);
 }
